refactor: Auth 타입 분리
